pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "koa-cli"
version = "0.3.0"
description = "CLI for submitting and managing KOA HPC jobs from a local machine."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
authors = [
  {name = "Michael Buriek"},
  {name = "Yosub Shin"}
]
dependencies = [
  "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "black",
  "ruff",
  "pytest>=7.0",
  "pytest-mock>=3.10",
]

[project.scripts]
koa = "koa_cli.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
